+++ com.oracle.max.vm.ext.maxri/src/com/oracle/max/vm/ext/maxri/MaxXirGenerator.java
@@ -32,8 +32,8 @@ import java.lang.reflect.*;
 import java.util.*;
 
 import com.oracle.max.asm.target.armv7.*;
-import com.sun.cri.ci.*;
 import com.sun.cri.ci.CiAddress.*;
+import com.sun.cri.ci.*;
 import com.sun.cri.ri.*;
 import com.sun.cri.ri.RiType.*;
 import com.sun.cri.xir.*;
@@ -54,7 +54,7 @@ import com.sun.max.vm.compiler.target.*;
 import com.sun.max.vm.heap.*;
 import com.sun.max.vm.heap.debug.*;
 import com.sun.max.vm.layout.*;
-import com.sun.max.vm.methodhandle.VMTarget;
+import com.sun.max.vm.methodhandle.*;
 import com.sun.max.vm.object.*;
 import com.sun.max.vm.runtime.*;
 import com.sun.max.vm.runtime.aarch64.*;
@@ -358,6 +358,14 @@ public class MaxXirGenerator implements RiXirGenerator {
             asm.stackOverflowCheck();
         }
 
+        if (MaxineVM.isRunning() &&
+                methodIsAllocationProfilerEntryOrExitPoint(method, CompilationBroker.AllocationProfilerEntryPoint)) {
+            XirOperand  tla             = asm.createRegisterTemp("TLA", WordUtil.archKind(), this.LATCH_REGISTER);
+            XirConstant offsetToProfile = asm.i(VmThreadLocal.PROFILER_TLA.offset);
+            XirConstant constantOne     = asm.i(1);
+            asm.pstore(WordUtil.archKind(), tla, offsetToProfile, constantOne, false);
+        }
+
         return new XirSnippet(finishTemplate(asm, "prologue"));
     }
 
@@ -375,7 +383,19 @@ public class MaxXirGenerator implements RiXirGenerator {
         if (callee.isTemplate()) {
             return null;
         }
+        if (MaxineVM.isRunning() &&
+                methodIsAllocationProfilerEntryOrExitPoint(method, CompilationBroker.AllocationProfilerExitPoint)) {
+            XirOperand  tla             = asm.createRegisterTemp("TLA", WordUtil.archKind(), this.LATCH_REGISTER);
+            XirConstant offsetToProfile = asm.i(VmThreadLocal.PROFILER_TLA.offset);
+            XirConstant constantZero    = asm.i(0);
+            asm.pstore(WordUtil.archKind(), tla, offsetToProfile, constantZero, false);
+        }
         return new XirSnippet(epilogueTemplate);
+    }
+
+    private boolean methodIsAllocationProfilerEntryOrExitPoint(RiResolvedMethod method, String entryOrExitPoint) {
+        final String methodFullName = method.toString();
+        return methodFullName.substring(0, methodFullName.indexOf('(')).equals(entryOrExitPoint);
     }
 
     @Override
@@ -1087,6 +1107,10 @@ public class MaxXirGenerator implements RiXirGenerator {
         Scale scale = Scale.fromInt(elemSize);
         alignArraySize(length, arraySize, elemSize, scale);
 
+        if (MaxineVM.profileThatObject()) {
+            callRuntimeThroughStub(asm, "callProfilerArray", null, arraySize, hub);
+        }
+
         asm.pload(WordUtil.archKind(), cell, etla, offsetToTLABMark, false);
         asm.pload(WordUtil.archKind(), tlabEnd, etla, offsetToTLABEnd, false);
         asm.add(newMark, cell, arraySize);
@@ -1137,8 +1161,13 @@ public class MaxXirGenerator implements RiXirGenerator {
         Scale scale = Scale.fromInt(elemSize);
         alignArraySize(length, arraySize, elemSize, scale);
 
+        if (MaxineVM.profileThatObject()) {
+            callRuntimeThroughStub(asm, "callProfilerArray", null, arraySize, hub);
+        }
+
         asm.pload(WordUtil.archKind(), cell, etla, offsetToTLABMark, false);
         asm.pload(WordUtil.archKind(), tlabEnd, etla, offsetToTLABEnd, false);
+
         asm.add(newMark, cell, arraySize);
         asm.jgt(slowPath, newMark, tlabEnd);
         asm.pstore(WordUtil.archKind(), etla, offsetToTLABMark, newMark, false);
@@ -1274,6 +1303,11 @@ public class MaxXirGenerator implements RiXirGenerator {
         asm.pload(WordUtil.archKind(), etla, tla, asm.i(VmThreadLocal.ETLA.offset), false);
         asm.pload(WordUtil.archKind(), cell, etla, offsetToTLABMark, false);
         asm.pload(WordUtil.archKind(), tlabEnd, etla, offsetToTLABEnd, false);
+
+        if (MaxineVM.profileThatObject()) {
+            callRuntimeThroughStub(asm, "callProfiler", null, tupleSize, hub);
+        }
+
         asm.add(newMark, cell, tupleSize);
         asm.jlteq(ok, newMark, tlabEnd);
         // Slow path.
@@ -1339,6 +1373,9 @@ public class MaxXirGenerator implements RiXirGenerator {
         asm.pload(WordUtil.archKind(), etla, tla, asm.i(VmThreadLocal.ETLA.offset), false);
         asm.pload(WordUtil.archKind(), cell, etla, offsetToTLABMark, false);
         asm.pload(WordUtil.archKind(), tlabEnd, etla, offsetToTLABEnd, false);
+        if (MaxineVM.profileThatObject()) {
+            callRuntimeThroughStub(asm, "callProfiler", null, tupleSize, hub);
+        }
         asm.add(newMark, cell, tupleSize);
         asm.jgt(slowPath, newMark, tlabEnd);
         asm.pstore(WordUtil.archKind(), etla, offsetToTLABMark, newMark, false);
@@ -1356,6 +1393,7 @@ public class MaxXirGenerator implements RiXirGenerator {
             asm.pstore(CiKind.Int, cell, asm.i(arrayLayout().arrayLengthOffset()), asm.i(hubFirstWordIndex()), false);
         }
         asm.mov(result, cell);
+
         asm.bindOutOfLine(slowPath);
         callRuntimeThroughStub(asm, "slowPathAllocate", cell, tupleSize, etla);
         asm.jmp(done);
@@ -2088,6 +2126,25 @@ public class MaxXirGenerator implements RiXirGenerator {
                 FatalError.check(vmConfig().heapScheme().usesTLAB(), "HeapScheme must use TLAB");
             }
             return ((HeapSchemeWithTLAB) vmConfig().heapScheme()).c1xSlowPathAllocate(Size.fromInt(size), etla);
+        }
+
+        /**
+         * Runtime entry point for object allocation profiling.
+         * @param size of the profiled object.
+         * @param hub object hub to obtain the type of the profiled object.
+         */
+        public static void callProfiler(int size, Hub hub) {
+            if (MaxineVM.isDebug()) {
+                FatalError.check(vmConfig().heapScheme().usesTLAB(), "HeapScheme must use TLAB");
+            }
+            ((HeapSchemeWithTLAB) vmConfig().heapScheme()).profile(size, hub);
+        }
+
+        public static void callProfilerArray(int size, Hub hub) {
+            if (MaxineVM.isDebug()) {
+                FatalError.check(vmConfig().heapScheme().usesTLAB(), "HeapScheme must use TLAB");
+            }
+            ((HeapSchemeWithTLAB) vmConfig().heapScheme()).profileArray(size, hub);
         }
 
         public static Pointer flushLog(Pointer logTail) {