+++ com.sun.max/src/com/sun/max/vm/thread/VmThread.java
@@ -239,7 +239,6 @@ public class VmThread {
      */
     public final int uuid;
 
-
     /**
      * Denotes if this thread was started as a daemon. This property is only set once a thread
      * is about to run (for a VM created thread) or is running (for an attached thread)