+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.7.0] - Unreleased
+
+### Added
+
+- Allocation Profiler (-XX:+AllocationProfilerAll -XX:+AllocationProfilerPrintHistogram) gathers stats about object 
+allocations
+- Fine grain allocation profiling with -XX:AllocationProfilerEntryPoint -XX:AllocationProfilerExitPoint
+- -XX:+LogCompiledMethods to print all runtime compiled methods in the application
+
 ## [2.6.0] - 2018-12-05
 
 ### Added